frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "dash>=3.2.0",
    "fastapi[standard]>=0.119.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
]
